pyproject.toml
[project]
name = "gpt-oss-eval"
version = "0.1.0"
description = "Testing OpenAI GPT OSS model via Groq API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
    "chainlit>=1.0.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-tavily>=0.1.0",
    "ddgs>=9.0.0",
    "langgraph>=0.2.0",
    "pillow>=11.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
